Make the dependency trees only check for the .rel attributes.
The dependency trees used to require that a word be part of the tree if it had
any "annotation attribute" that made them require just about everything. Now a
tree requires a word to be part of the tree if and only if the @...rel attribute
that corresponds to the tree is set.